rename: domain/lecture/enrollment 패키지를 제거하고 패키지 정리
- enrollment가 lecture 도메인에 속한다고 판
