+++ src/main/java/com/leoh/hhweek2/domain/lecture/Lecture.java
@@ -1,7 +1,6 @@
 package com.leoh.hhweek2.domain.lecture;
 
 import com.leoh.hhweek2.domain.common.BaseEntity;
-import com.leoh.hhweek2.domain.lecture.enrollment.Enrollment;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
