+++ src/main/java/com/leoh/hhweek2/infrastructures/core/enrollment/EnrollmentJpaRepository.java
@@ -1,6 +1,6 @@
 package com.leoh.hhweek2.infrastructures.core.enrollment;
 
-import com.leoh.hhweek2.domain.lecture.enrollment.Enrollment;
+import com.leoh.hhweek2.domain.lecture.Enrollment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
