--- src/main/java/com/leoh/hhweek2/interfaces/api/lecture/EnrollRequest.java
@@ -1,4 +1,0 @@
-package com.leoh.hhweek2.interfaces.api.lecture;
-
-public record EnrollRequest(long userId) {
-}
