+++ src/main/java/com/leoh/hhweek2/interfaces/api/lecture/LectureController.java
@@ -1,6 +1,7 @@
 package com.leoh.hhweek2.interfaces.api.lecture;
 
 import com.leoh.hhweek2.application.lecture.LectureFacade;
+import com.leoh.hhweek2.interfaces.api.lecture.dto.*;
 import com.leoh.hhweek2.support.http.response.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
