+++ src/main/java/com/leoh/hhweek2/interfaces/api/lecture/dto/EnrollRequest.java
@@ -1,0 +1,4 @@
+package com.leoh.hhweek2.interfaces.api.lecture.dto;
+
+public record EnrollRequest(long userId) {
+}
