+++ src/main/java/com/leoh/hhweek2/interfaces/api/lecture/dto/EnrollResponse.java
@@ -1,7 +1,7 @@
-package com.leoh.hhweek2.interfaces.api.lecture;
+package com.leoh.hhweek2.interfaces.api.lecture.dto;
 
-import com.leoh.hhweek2.domain.lecture.LectureServiceResponse;
-import com.leoh.hhweek2.domain.lecture.enrollment.EnrollmentServiceResponse;
+import com.leoh.hhweek2.domain.lecture.dto.LectureServiceResponse;
+import com.leoh.hhweek2.domain.lecture.dto.EnrollmentServiceResponse;
 
 public record EnrollResponse(
         long lectureId,
